third_party/blink/web_tests/flag-specific/highdpi/accessibility/inline-text-bounds-for-range-expected.txt

This is a testharness.js-based test.
FAIL Tests that we can compute the bounds of a range of text from the accessibility tree. assert_approx_equals: expected 8 +/- 1 but got 12
Harness: the test ran to completion.
